LayoutTests/imported/w3c/web-platform-tests/2dcontext/drawing-rectangles-to-the-canvas/2d.strokeRect.nonfinite-expected.txt

2d.strokeRect.nonfinite
strokeRect() with Infinity/NaN is ignored
Actual output:
Expected output:


PASS strokeRect() with Infinity/NaN is ignored 
